create Author
 createBook => createInput
 (handle create book AND author in the same component)
